requirements.txt
# Core
langchain==0.1.13
langchain-community==0.0.34
unstructured==0.11.1
PyPDF2==3.0.1
pypdf==4.2.0
pdfminer.six==20221105
tiktoken==0.6.0
tqdm==4.66.4
matplotlib==3.8.0
# NLP + OCR
nltk==3.8.1
pandas==2.2.2
opencv-python-headless==4.9.0.80
unstructured_pytesseract==0.3.12
unstructured_inference==0.7.26
pdf2image==1.17.0

# Optional: if using sentence transformers or embedding models
sentence-transformers==2.2.2
huggingface_hub==0.16.4
faiss-cpu==1.7.4
python-docx==1.1.0
openai
